hdr: add BaseStoryPart.new_pic_inline()
This is essentially migrated "up" from DocumentPart, which now
subclasses BaseStoryPart.
