+++ docx/parts/document.py
@@ -6,17 +6,16 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from docx.document import Document
 from docx.opc.constants import RELATIONSHIP_TYPE as RT
-from docx.opc.part import XmlPart
-from docx.oxml.shape import CT_Inline
 from docx.parts.hdrftr import FooterPart, HeaderPart
 from docx.parts.numbering import NumberingPart
 from docx.parts.settings import SettingsPart
+from docx.parts.story import BaseStoryPart
 from docx.parts.styles import StylesPart
 from docx.shape import InlineShapes
 from docx.shared import lazyproperty
 
 
-class DocumentPart(XmlPart):
+class DocumentPart(BaseStoryPart):
     """Main document part of a WordprocessingML (WML) package, aka a .docx file.
 
     Acts as broker to other parts such as image, core properties, and style parts. It
@@ -103,17 +102,6 @@ class DocumentPart(XmlPart):
         document.
         """
         return InlineShapes(self._element.body, self)
-
-    def new_pic_inline(self, image_descriptor, width, height):
-        """Return a newly-created `w:inline` element.
-
-        The element contains the image specified by *image_descriptor* and is scaled
-        based on the values of *width* and *height*.
-        """
-        rId, image = self.get_or_add_image(image_descriptor)
-        cx, cy = image.scaled_dimensions(width, height)
-        shape_id, filename = self.next_id, image.filename
-        return CT_Inline.new_pic_inline(shape_id, rId, filename, cx, cy)
 
     @property
     def next_id(self):
