+++ tests/parts/test_story.py
@@ -7,12 +7,14 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import pytest
 
 from docx.enum.style import WD_STYLE_TYPE
+from docx.image.image import Image
 from docx.package import Package
 from docx.parts.document import DocumentPart
 from docx.parts.story import BaseStoryPart
 from docx.styles.style import BaseStyle
 
-from ..unitutil.mock import instance_mock, property_mock
+from ..unitutil.file import snippet_text
+from ..unitutil.mock import instance_mock, method_mock, property_mock
 
 
 class DescribeBaseStoryPart(object):
@@ -44,6 +46,20 @@ class DescribeBaseStoryPart(object):
         document_part_.get_style_id.assert_called_once_with(style_, style_type)
         assert style_id == "BodyText"
 
+    def it_can_create_a_new_pic_inline(self, get_or_add_image_, image_, next_id_prop_):
+        get_or_add_image_.return_value = "rId42", image_
+        image_.scaled_dimensions.return_value = 444, 888
+        image_.filename = "bar.png"
+        next_id_prop_.return_value = 24
+        expected_xml = snippet_text("inline")
+        story_part = BaseStoryPart(None, None, None, None)
+
+        inline = story_part.new_pic_inline("foo/bar.png", width=100, height=200)
+
+        get_or_add_image_.assert_called_once_with(story_part, "foo/bar.png")
+        image_.scaled_dimensions.assert_called_once_with(100, 200)
+        assert inline.xml == expected_xml
+
     def it_knows_the_main_document_part_to_help(self, package_, document_part_):
         package_.main_document_part = document_part_
         story_part = BaseStoryPart(None, None, None, package_)
@@ -61,6 +77,18 @@ class DescribeBaseStoryPart(object):
     @pytest.fixture
     def _document_part_prop_(self, request):
         return property_mock(request, BaseStoryPart, "_document_part")
+
+    @pytest.fixture
+    def get_or_add_image_(self, request):
+        return method_mock(request, BaseStoryPart, "get_or_add_image")
+
+    @pytest.fixture
+    def image_(self, request):
+        return instance_mock(request, Image)
+
+    @pytest.fixture
+    def next_id_prop_(self, request):
+        return property_mock(request, BaseStoryPart, "next_id")
 
     @pytest.fixture
     def package_(self, request):
